requirements.txt
aiogram==2.23.1
python-dotenv==0.21.0
mysql-connector-python==8.0.31
passlib==1.7.4